Refactor profile handling into manager
